Refactor: move styled components to separate files
