+++ client/src/components/Header.jsx
@@ -1,21 +1,6 @@
 import React from 'react';
-import styled from 'styled-components';
 
-const Heading = styled.header`
-  align-items: center;
-  display: flex;
-  height: 10vh;
-  justify-content: center;
-  width: 100%;
-  `;
-
-const Title = styled.h1`
-  color: ${props=>props.theme.colors.ternary};
-  font-family: Notable;
-  font-size: var(--font-xl);
-  text-align: center;
-  text-shadow: 1px 1px 1px rgba(249, 248, 113, 0.6);
-`;
+import { Heading, Title } from './styled/HeaderStyle';
 
 // CLICKING ON HAMBURGER OPENS USER PANEL
