+++ client/src/components/MainPage.jsx
@@ -1,21 +1,11 @@
 import React from 'react';
-import styled from 'styled-components';
 
 import SearchBar from './main-page/SearchBar';
 import SelectBar from './main-page/SelectBar';
-import Slider from './main-page/slider/Slider';
+import Slider from './main-page/Slider';
 import SelectedMovie from './main-page/SelectedMovie';
 
-const MainPageContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  height: 90vh;
-  justify-content: space-between;
-  margin: 0 auto;
-  max-height: 600px;
-  max-width: 700px;
-`;
-
+import MainPageContainer from './styled/MainPageStyle';
 
 class MainPage extends React.Component {
   render() {
@@ -25,7 +15,7 @@ class MainPage extends React.Component {
           text="find movie"
           />
         <SelectBar
-          active={ this.props.sortBy }
+          sortBy={ this.props.sortBy }
           onSortSelect={ this.props.onSortSelect }
           />
         <Slider
