+++ client/src/components/main-page/SearchBar.jsx
@@ -1,24 +1,6 @@
 import React from 'react';
-import styled from 'styled-components';
 
-import Bar from './Bar';
-
-const SearchInput = styled.input`
-  background: transparent;
-  border: none;
-  color: ${props=>props.theme.colors.main}
-  cursor: text;
-  font: inherit;
-  outline: none;
-  &::placeholder {
-    color: ${props=>props.theme.colors.main}
-  }
-`;
-
-const SearchIcon = styled.i`
-  color: ${props => props.theme.colors.secondary };
-`;
-
+import { Bar, SearchInput, SearchIcon } from './styled/SearchBarStyle';
 
 const SearchBar = props => {
   return (
