+++ client/src/components/main-page/slider/Slide.jsx
@@ -1,13 +1,6 @@
 import React from 'react';
-import styled from 'styled-components';
 
-const SlideImg = styled.img`
-  box-shadow: 0 0 8px 1px rgba(196, 196, 196, 0.2);
-  cursor: pointer;
-  height: 118px;
-  margin: 0 15px;
-  min-width: 80px;
-`;
+import SlideImg from './styled/SlideStyle';
 
 const Slide = props => <SlideImg
   src={props.image}
