+++ client/src/components/main-page/slider/SliderArrows.jsx
@@ -1,13 +1,6 @@
 import React from 'react';
-import styled from 'styled-components';
 
-const Arrow = styled.i`
-  color: ${props => props.theme.colors.secondaryTrans};
-  z-index: 10;
-  &:hover {
-    color: ${props => props.theme.colors.secondary};
-  }
-`;
+import Arrow from './styled/SlideArrowsStyle';
 
 const RightArrow = props =>
   <Arrow
