+++ client/src/components/main-page/slider/styled/SlideArrowsStyle.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const Arrow = styled.i`
+  color: ${props => props.theme.colors.secondaryTrans};
+  z-index: 10;
+  &:hover {
+    color: ${props => props.theme.colors.secondary};
+  }
+`;
+
+export default Arrow;
