+++ client/src/components/main-page/slider/styled/SlideStyle.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const SlideImg = styled.img`
+  box-shadow: 0 0 8px 1px rgba(196, 196, 196, 0.2);
+  cursor: pointer;
+  height: 118px;
+  margin: 0 15px;
+  min-width: 80px;
+`;
+
+export default SlideImg;
