+++ client/src/components/styled/MainPageStyle.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const MainPageContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  height: 90vh;
+  justify-content: space-between;
+  margin: 0 auto;
+  max-height: 600px;
+  max-width: 700px;
+`;
+
+export default MainPageContainer;
